+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Change Log
 All notable changes to the "Diff Folders" extension will be documented in this file.
 
-## [1.0.4] - 2023-01-22
+## [1.1.0] - 2023-02-12
+
+### Added
+
+- Added `l13Diff.ignoreUTF8BOM` to ignore UTF-8 BOM in text files.
 
 ### Fixed
 
-- Fixed symlink copy if file or folder does not exist [Issue #104](https://github.com/L13/vscode-diff/issues/104)
-- Fixed JSONC parser for trailing comma in an objects or array.
+- Fixed copy symlink if file or folder does not exist [Issue #104](https://github.com/L13/vscode-diff/issues/104)
+- Fixed JSONC parser for trailing comma in an object or array.
+
+## [1.0.3] - 2021-06-23
+
+### Changed
+
+- Changed `extensionKind` [Issue #98](https://github.com/L13/vscode-diff/issues/98)
 
 ## [1.0.2] - 2021-06-13